Avoid printing anything when autothrottle is no-op
Whenever there is no change to the throttle rate, autothrottle was still
saying "Proposed throttle is within 0.00% of the previous throttle"
message and the "Replication throttle rate for broker....", which was a
bit misleading because in that case, autothrottle was actually changing
nothing to the system state.

Makes it silent when nothing change.